requirements.txt
yt-dlp
curl-cffi
openai-whisper
google-genai
python-dotenv
edge-tts
aiogram
